fix pt1
